Use asset_packager for admin side.
It's weird, right? We just have one stylesheet, so why bother?

The problem is sass and heroku: I need to pre-generate a css file that
will get checked into version control, and asset_packager is how I do
that for the main site. Might as well be consistent, rather than not
.gitignore-ing admin.css.